shifted patch summary logging to framework plugin and patch id renames
