+++ plugins/framework/src/fms_acceleration/framework_plugin.py
@@ -14,7 +14,7 @@
 
 # Standard
 from dataclasses import dataclass
-from typing import Any, Dict, List, Optional, Set, Tuple
+from typing import Any, Dict, List, Optional, Set, Tuple, Callable
 import importlib
 import sys
 
@@ -23,6 +23,29 @@ from accelerate import Accelerator
 from peft import LoraConfig
 from transformers import TrainingArguments
 import torch
+
+from transformers.utils import logging
+
+# want to use the transformers logger, but a bit of pain
+logger = logging.get_logger(__name__)  # pylint: disable=invalid-name
+logger.setLevel(logging._get_default_logging_level())
+logger.addHandler(logging._default_handler)
+
+def log_patch_summary(
+    logging_func: Callable = None,
+):
+    if logging_func is None:
+        logging_func = print
+
+    # this is a guarded import, because the model rule registration
+    # does not need to be loaded unless patch_model is required
+    # Local
+    from fms_acceleration.model_patcher import (  # pylint: disable=import-outside-toplevel
+        patch_model_summary,
+    )
+
+    for line in patch_model_summary().split("\n"):
+        logging_func(line)
 
 
 @dataclass
@@ -149,6 +172,11 @@ class AccelerationPlugin:
         # Finally apply all registered patches to the model
         from .model_patcher import ModelPatcher # pylint: disable=import-outside-toplevel
         ModelPatcher.patch(model)
+
+        # if patching is done, print patch summary to logger
+        if len(ModelPatcher.history)>0:
+            log_patch_summary(logging_func=logger.info)
+
         return []
 
     def _check_config_and_maybe_check_values(
